requirements.txt
assemblyai==0.27.0
certifi==2024.6.2
pyperclip==1.9.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
python-dotenv==1.0.1
pytube==15.0.0
